added styles, separated them in on .css file for each .js file, and included them all in a styles folder inside the src folder.
