+++ src/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import './styles/App.css';
 import Header from './Header';
 import Recipes from './Recipes';
 import Footer from './Footer';
