+++ src/Header.js
@@ -5,7 +5,7 @@ const Header = () => {
     return (
         <header>
             <div class="wrapper">
-                <h1><span class="headerSpan">Choose your preferred cuisine:</span></h1>
+                <h1><span class="headerSpan">Cuisine-based recipes!</span></h1>
             </div>
         </header>
     )
