+++ src/styles/Form.css
@@ -1,0 +1,44 @@
+form {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin-top: -240px;
+    /* background-color: rgb(215, 215, 239); */
+    /* width: 50%; */
+}
+
+.formLabel {
+    margin: 0px 50px 50px ;
+    background-color: rgb(215, 215, 239);
+    padding: 5px;
+    font-size: 20px;
+    width: 550px; 
+}
+
+select, button {
+    background-color: rgb(237, 234, 233);
+    font-size: 20px;
+}
+
+select {
+    width: 300px;
+    
+}
+
+button {
+    margin: 20px 
+    10px;
+    padding: 5px 30px;
+}
+
+button:hover {
+    background-color: rgb(237, 208, 199);
+    transition: .5s all ease;
+}
+
+.selectionButton {
+    max-width: 500px;
+    text-align: center;
+    padding: 5px;
+}
