+++ src/styles/Recipes.css
@@ -1,0 +1,22 @@
+.photos {
+    list-style-type: none;
+}
+
+.recipesContainer {
+    display: flex;
+    justify-content: left;
+    flex-flow: row-reverse;
+    align-items: center;
+    /* margin-top: 80px; */
+}
+
+.photos {
+    margin-top: 100px;
+}
+
+.recipesContainer img {
+    border-radius: 50%;
+    border: 5px solid white;
+    margin: 20px 80px;
+}
+
